move neon code to speck_neon.c & support android build
